+++ src/main/java/ctlab/bn/action/Cache.java
@@ -2,11 +2,11 @@ package ctlab.bn.action;
 
 public interface Cache {
     void disable(short action);
-    void reEnable(short action, float ll);
+    void reEnable(short action, double ll);
     boolean contains(short action);
-    float loglikelihood();
+    double loglikelihood();
     Short randomAction();
-    float min();
-    Short add(short action, float ll);
+    double min();
+    Short add(short action, double ll);
     boolean isFull();
 }
